--- src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/factory/BrAPIState.java
@@ -1,4 +1,0 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.factory;
-
-public interface BrAPIState {
-}
