+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/ExpUnitMiddleware.java
@@ -1,6 +1,6 @@
 package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware;
 
-import org.breedinginsight.brapps.importer.services.processors.experiment.middleware.Middleware;
+import org.breedinginsight.brapps.importer.services.processors.experiment.model.middleware.Middleware;
 import org.breedinginsight.brapps.importer.services.processors.experiment.model.ExpUnitMiddlewareContext;
 
 /**
