+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/factory/BrAPIState.java
@@ -1,0 +1,4 @@
+package org.breedinginsight.brapps.importer.services.processors.experiment.factory;
+
+public interface BrAPIState {
+}
