+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/factory/action/WorkflowReadInitialization.java
@@ -1,10 +1,11 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.factory;
+package org.breedinginsight.brapps.importer.services.processors.experiment.factory.action;
 
 import io.micronaut.context.annotation.Prototype;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.brapi.client.v2.model.exceptions.ApiException;
-import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.entity.ExperimentImportEntity;
+import org.breedinginsight.brapps.importer.services.processors.experiment.factory.BrAPIState;
+import org.breedinginsight.brapps.importer.services.processors.experiment.factory.entity.ExperimentImportEntity;
 import org.breedinginsight.utilities.Utilities;
 
 import java.util.List;
