+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/factory/entity/PendingGermplasm.java
@@ -1,4 +1,4 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.entity;
+package org.breedinginsight.brapps.importer.services.processors.experiment.factory.entity;
 
 import io.micronaut.context.annotation.Prototype;
 import org.brapi.client.v2.model.exceptions.ApiException;
@@ -15,7 +15,6 @@ import org.breedinginsight.services.exceptions.DoesNotExistException;
 import org.breedinginsight.services.exceptions.MissingRequiredInfoException;
 import org.breedinginsight.services.exceptions.UnprocessableEntityException;
 
-import javax.inject.Inject;
 import java.util.*;
 import java.util.stream.Collectors;
